Omit inner classes that are not class members.
If one declares a *named* class in a method then
it's reported as an inner class but not as
a member of outer class defining a method. We
can detect such case by looking at outer class
record in InnerClasses table.

We exclude such classes from import considerations
as they should not be allowed to be imported.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>