Blacklist a few classes.
There are a few problematic classes in scala-library
that do not matter for this test but make it fail.

Blacklisting them due to lack of better idea how
to fix the real problem of detecting classes that
one should be able to import from Java and ones
that have no hope to be imported from Java.